Update performance tests
